+++ java/core/src/test/java/co/worklytics/psoxy/rules/workdata/WorkDataGenericTests.java
@@ -62,6 +62,7 @@ public class WorkDataGenericTests {
         Container.ForConfigService.class,
         MockModules.ForSecretStore.class,
         MockModules.ForHostEnvironment.class,
+        MockModules.ForResourceService.class,
         TestModules.ForProxyConstants.class,
     })
     public interface Container {